# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FDA Regulatory Intelligence Platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/chat": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Chat",
        "description": "Run the guarded graph and persist one authoritative final turn.",
        "operationId": "chat_api_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or token invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Caller is not authorized for this operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible to the caller.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "A required service is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat/stream": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Chat Stream",
        "description": "Stream node progress and one durable committed/refused final event.",
        "operationId": "chat_stream_api_chat_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Named SSE events carrying the documented chat stream protocol.",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ChatStreamEvent"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or token invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Caller is not authorized for this operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible to the caller.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "A required service is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Answer": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text",
            "description": "Generated answer text with inline citation markers."
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceCard"
            },
            "type": "array",
            "title": "Evidence",
            "description": "Evidence Cards backing each cited claim."
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "description": "Overall answer confidence from guardrails."
          },
          "refused": {
            "type": "boolean",
            "title": "Refused",
            "description": "True when the workflow refused due to insufficient grounding.",
            "default": false
          },
          "refusal_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refusal Reason",
            "description": "Explanation when the answer was refused."
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id",
            "description": "Owning session id."
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At",
            "description": "Answer generation timestamp."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text",
          "evidence",
          "confidence"
        ],
        "title": "Answer",
        "description": "A generated answer with mandatory supporting evidence.\n\nEvery answer must include at least one :class:`EvidenceCard`. When the\nworkflow cannot ground its claims, ``refused`` is set and ``text`` explains\nthe refusal instead of asserting an unsupported answer."
      },
      "AnswerRequest": {
        "properties": {
          "query": {
            "type": "string",
            "minLength": 1,
            "title": "Query",
            "description": "Natural-language regulatory question."
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id",
            "description": "Conversation/session identifier for continuity."
          },
          "filters": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Filters",
            "description": "Optional retrieval filters (e.g. status, office)."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "query"
        ],
        "title": "AnswerRequest",
        "description": "A user query submitted to the agent workflow."
      },
...
      "ChatResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "description": "Server-owned chat session identifier."
          },
          "turn_id": {
            "type": "string",
            "title": "Turn Id",
            "description": "Durable final turn identifier."
          },
          "answer": {
            "$ref": "#/components/schemas/Answer",
            "description": "Citation-first answer or refusal."
          },
          "transparency": {
            "$ref": "#/components/schemas/TransparencyMetadata",
            "description": "Node/tool execution transparency."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "session_id",
          "turn_id",
          "answer",
          "transparency"
        ],
        "title": "ChatResponse",
        "description": "Non-streaming chat response with transparency metadata."
      },
      "ChatStreamEvent": {
        "properties": {
          "event": {
            "type": "string",
            "enum": [
              "started",
              "node",
              "committed",
              "refused",
              "error",
              "done"
            ],
            "title": "Event"
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "event"
        ],
        "title": "ChatStreamEvent",
        "description": "A single SSE lifecycle event for streaming chat.\n\n``token`` events are intentionally not modeled here: the current LangGraph\nworkflow only streams at node boundaries, not token-by-token, so ``node``\nevents are the finest-grained preview available until Phase 4 wires\n``LLMClient.stream`` into ``generate_answer``."
      },
...
      "DocumentStatus": {
        "type": "string",
        "enum": [
          "Draft",
          "Final",
          "Withdrawn"
        ],
        "title": "DocumentStatus",
        "description": "Lifecycle status of an FDA guidance document."
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stable machine-readable error code."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable, secret-free error message."
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id",
            "description": "Correlation id for support/debugging."
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details",
            "description": "Optional field-level validation details."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "description": "Sanitized error detail returned to API clients."
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "Envelope for all structured API error responses."
      },
      "EvidenceCard": {
        "properties": {
          "citation_id": {
            "type": "string",
            "title": "Citation Id",
            "description": "Stable answer-local citation marker, such as '[1]'."
          },
          "document_id": {
            "type": "string",
            "title": "Document Id",
            "description": "Stable internal FDA document identifier."
          },
          "chunk_id": {
            "type": "string",
            "title": "Chunk Id",
            "description": "Stable source chunk identifier for exact passage lookup."
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "FDA guidance title."
          },
          "section_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section Id",
            "description": "Normalized section identifier when available."
          },
          "section_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section Title",
            "description": "Human-readable section heading."
          },
          "page_number": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Number",
            "description": "PDF page number for source inspection."
          },
          "passage": {
            "type": "string",
            "title": "Passage",
            "description": "Exact quoted or cited source passage."
          },
          "source_url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Source Url",
            "description": "FDA source URL."
          },
          "version_hash": {
            "type": "string",
            "title": "Version Hash",
            "description": "SHA-256 hash of the source document version."
          },
          "document_status": {
            "$ref": "#/components/schemas/DocumentStatus",
            "description": "Draft, Final, or Withdrawn."
          },
          "retrieval_score": {
            "type": "number",
            "title": "Retrieval Score",
            "description": "Dense/BM25/RRF retrieval score used for ranking."
          },
          "rerank_score": {
            "type": "number",
            "title": "Rerank Score",
            "description": "Cross-encoder reranker score."
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "description": "Final evidence confidence used by guardrails."
          },
          "retrieved_at": {
            "type": "string",
            "format": "date-time",
            "title": "Retrieved At",
            "description": "Timestamp of retrieval for auditability."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "citation_id",
          "document_id",
          "chunk_id",
          "title",
          "passage",
          "source_url",
          "version_hash",
          "document_status",
          "retrieval_score",
          "rerank_score",
          "confidence"
        ],
        "title": "EvidenceCard",
        "description": "A verifiable citation binding a claim to an FDA source passage."
      },
...
      "TransparencyMetadata": {
        "properties": {
          "execution_trace": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Execution Trace",
            "description": "Ordered workflow node execution trace."
          },
          "retrieval_diagnostics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Retrieval Diagnostics",
            "description": "Dense/BM25/reranker diagnostics from retrieval."
          },
          "provider_fallback_trace": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Provider Fallback Trace",
            "description": "LLM provider fallback events, if any."
          },
          "ignored_filter_keys": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Ignored Filter Keys",
            "description": "Unrecognized filter keys ignored during routing."
          },
          "guardrail_errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Guardrail Errors",
            "description": "Guardrail failures encountered during the workflow."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "TransparencyMetadata",
        "description": "Node/tool execution transparency surfaced to API clients."
      },
...
    }
  }
}